REF: move abi_tag and python_tag guessers into separate module.
